app/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
fastapi==0.103.2
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
httptools==0.6.1
idna==3.4
install==1.3.5
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
pydantic==1.10.13
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
watchfiles==0.21.0
websockets==11.0.3
yarl==1.9.2
